refactor: extract reusable access profile dialog
